backend/requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
sqlalchemy>=1.4.23,<1.5.0
pydantic>=1.8.0,<2.0.0
python-jose[cryptography]>=3.3.0,<3.4.0
passlib[bcrypt]>=1.7.4,<1.8.0
python-multipart>=0.0.5,<0.0.6
alembic>=1.7.3,<1.8.0
psycopg2-binary>=2.9.1,<2.10.0
requests>=2.26.0,<2.27.0
pytest>=6.2.5,<6.3.0
httpx>=0.19.0,<0.20.0
pyotp>=2.6.0,<2.7.0
qrcode>=7.3.1,<7.4.0
pillow>=8.3.2,<8.4.0
cryptography>=35.0.0,<36.0.0
itsdangerous>=2.0.1,<2.1.0
starlette>=0.14.2,<0.15.0
aiofiles>=0.7.0,<0.8.0
python-dotenv>=0.19.0,<0.20.0